+++ workoutpact/LoginViewController.swift
@@ -27,8 +27,9 @@ class LoginViewController: UIViewController {
             // ensure that presentViewController happens from the main thread/queue
             dispatch_async(dispatch_get_main_queue(), {
                 if let storyboard = self.storyboard {
-                    let controller = storyboard.instantiateViewControllerWithIdentifier("TwoFactorViewController") as! TwoFactorViewController
-                    self.presentViewController(controller, animated: true, completion: nil)
+                    if let controller = storyboard.instantiateViewControllerWithIdentifier("TwoFactorViewController") as? TwoFactorViewController {
+                        self.presentViewController(controller, animated: true, completion: nil)
+                    }
                 }
             });